Cherry pick the folder organization commit from a different branch
