+++ packages/experiments/src/components/fluent/examplePages/examples/FluentStyles.Checkbox.Example.tsx
@@ -63,12 +63,7 @@ export class FluentStylesCheckboxExample extends React.Component<{}, ICheckboxBa
           styles={checkboxStyles}
         />
 
-        <Checkbox
-          label="Controlled checkbox"
-          checked={isChecked}
-          onChange={this._onControlledCheckboxChange}
-          styles={checkboxStyles}
-        />
+        <Checkbox label="Controlled checkbox" checked={isChecked} onChange={this._onControlledCheckboxChange} styles={checkboxStyles} />
 
         <Checkbox label="Checkbox rendered with boxSide &quot;end&quot;" boxSide="end" styles={checkboxStyles} />
