+++ packages/experiments/src/components/fluent/examplePages/examples/FluentStyles.Dialog.Example.tsx
@@ -27,8 +27,7 @@ export class FluentStylesDialogExample extends React.Component<{}, IFluentThemeD
           dialogContentProps={{
             type: DialogType.normal,
             title: 'All emails together',
-            subText:
-              'Your Inbox has changed. No longer does it include favorites, it is a singular destination for your emails.'
+            subText: 'Your Inbox has changed. No longer does it include favorites, it is a singular destination for your emails.'
           }}
           modalProps={{
             titleAriaId: 'myLabelId',
