+++ packages/experiments/src/components/fluent/examplePages/examples/FluentStyles.Link.Example.tsx
@@ -8,11 +8,11 @@ export class FluentStylesLinkExample extends React.Component<{}, {}> {
         <span>When a link has an href, </span>
         <Link href="http://dev.office.com/fabric/components/link">it renders as an anchor tag.</Link>
         <span> Without an href, </span>
-        <Link>the link is rendered as a button</Link>.
-        <span> You can also use the disabled attribute to create a </span>
+        <Link>the link is rendered as a button</Link>.<span> You can also use the disabled attribute to create a </span>
         <Link disabled={true} href="http://dev.office.com/fabric/components/link">
           disabled link
-        </Link>.
+        </Link>
+        .
       </div>
     );
   }
