+++ packages/experiments/src/components/fluent/examplePages/examples/FluentStyles.Slider.Example.tsx
