+++ packages/experiments/src/components/fluent/styles/CompoundButton.styles.ts
@@ -1,0 +1,7 @@
+import { fluentBorderRadius } from './styleConstants';
+
+export const CompoundButtonStyles = {
+  root: {
+    borderRadius: fluentBorderRadius
+  }
+};
