+++ packages/experiments/src/components/fluent/styles/Label.styles.ts
@@ -1,0 +1,7 @@
+import { FontWeights } from 'office-ui-fabric-react/lib/Styling';
+
+export const LabelStyles = {
+  root: {
+    fontWeight: FontWeights.semibold
+  }
+};
