+++ packages/experiments/src/components/fluent/styles/TextField.styles.ts
@@ -1,0 +1,7 @@
+import { fluentBorderRadius } from './styleConstants';
+
+export const TextFieldStyles = {
+  fieldGroup: {
+    borderRadius: fluentBorderRadius
+  }
+};
