+++ packages/experiments/src/components/fluent/styles/Toggle.styles.ts
@@ -1,0 +1,14 @@
+export const ToggleStyles = {
+  pill: {
+    width: '40px',
+    height: '20px',
+    borderRadius: '10px',
+    padding: '0 4px'
+  },
+  thumb: {
+    width: '12px',
+    height: '12px',
+    borderRadius: '12px',
+    borderColor: 'transparent'
+  }
+};
