+++ packages/experiments/src/components/fluent/styles/styleConstants.ts
@@ -1,0 +1,1 @@
+export const fluentBorderRadius = '2px';
