+++ packages/experiments/src/demo/AppDefinition.tsx
@@ -119,13 +119,13 @@ export const AppDefinition: IAppDefinition = {
           url: '#/examples/staticlist'
         },
         {
-          component: require<any>('../components/fluent/theme/FluentThemePage').FluentThemePage,
+          component: require<any>('../components/fluent/examplePages/FluentThemePage').FluentThemePage,
           key: 'FluentColorTheme',
           name: 'Fluent Color Theme',
           url: '#/examples/fluent-color-theme'
         },
         {
-          component: require<any>('../components/fluent/styles/FluentStylesPage').FluentStylesPage,
+          component: require<any>('../components/fluent/examplePages/FluentStylesPage').FluentStylesPage,
           key: 'FluentStyles',
           name: 'Fluent Component Styles',
           url: '#/examples/fluent-styles'
